requirements.txt
django-rest-framework
coreapi